Assistant checkpoint: Add dark theme toggle with sun/moon icons
Assistant generated file changes:
- _includes/navigation.html: Add theme toggle button
- _sass/_main.scss: Add dark theme styles, Update elements to use CSS variables
- assets/js/theme-switch.js: Add theme switch functionality
- _layouts/default.html: Add theme switch script

---

User prompt:

I want to add a dark theme for my blog site. It should be toggled using a small button with changes between sun and the moon icon. The dark theme and light theme should affect all the elements, like link highlighting. Nothing should mess up because of the theme switching. Make sure you accommodate all CSS changes that are needed to implement the dark theme. Implement a toggle switch. It should be on the top right of the website with a small sun and moon icon. When you click on it, it goes between light and dark theme. The default theme should be the Light theme.